fix(pr): roll back partial SHIPPED on structured ship failure (#860)
`pr create --sync` could leave Ticket.state == SHIPPED with no push
and no PR when ShipExecutor.run() returned a non-raising precondition
failure (RunnerResult(ok=False): no code host configured / no worktree
on ticket / branch already merged into base).

#838 made the sync ship atomic only for the *exception* path. A
RunnerResult(ok=False) does not raise, so execute_ship returned a
normal {"ok": False} dict, its transaction.atomic() committed, and
the outer _ship_sync transaction committed too -- persisting the
ship() FSM advance with no push and no PR (an unrunnable partial
state). The fix re-raises a failing execute_ship result inside the
shared atomic block as _ShipExecutionError carrying the real detail,
so both the raised and the ok=False paths share one rollback +
structured-surfacing path.

The ship-execution concern (_do_ship_transition, _enqueue_ship,
_ship_sync + the new error/abort helpers and their TypedDicts) is
extracted into a _ship_exec.py sibling, mirroring the existing
_ship_fsm.py / _pr_preview.py split, so pr.py stays within the
module-health LOC bar; pr.py re-exports the public names.

Regression test drives the structured-failure path through the real
(un-mocked) execute_ship task and asserts the FSM is NOT left at
SHIPPED and the real cause is surfaced.

Relates-to #860
