+++ BLUEPRINT.md
@@ -270,7 +270,7 @@ The central entity. One ticket per unit of work (maps to an issue/task in the tr
 
 **Worker enqueue pattern (BLUEPRINT §4 invariant):** transitions that own long I/O follow one rule — body stays pure (state change + metadata only), then `transaction.on_commit(lambda: execute_X.enqueue(self.pk))`. The state change and the queued work land atomically. Workers take a row lock (`select_for_update()`), re-check the source state, run the runner, and on success call the next transition. At-least-once delivery is safe because the state guard makes redelivery a no-op. See `teatree/core/runners/` for the runner classes and `teatree/core/tasks.py` for the workers.
 
-**Synchronous ship atomicity (`pr create --sync`, #838).** The inline path (`_ship_sync`) runs the `ship()` FSM transition **and** the inline `execute_ship` inside a *single* `transaction.atomic()` block. A `ShipExecutor.run()` exception — a `git push` precondition failure surfaces as a `CommandFailedError` — then rolls the `ship()` advance back, so the ship is all-or-nothing: either pushed + PR opened + FSM advanced, or the FSM is left untouched (safely re-runnable from `REVIEWED`). The exception is caught and returned as a structured `ShipExecuted` (`ok=False`, real cause in `detail`); pre-#838 it propagated unhandled, committing a partial `SHIPPED` (no push/PR) and crashing the `manage.py` subprocess so the CLI wrapper surfaced only an opaque `rc=1` with the real cause lost. This is the synchronous analogue of the worker enqueue pattern's "state change and queued work land atomically" invariant — the async path achieves it via `on_commit`, the sync path via the shared transaction. The block's `on_commit` enqueue fires only on commit (success); `execute_ship`'s own state guard makes a later worker pickup a no-op.
+**Synchronous ship atomicity (`pr create --sync`, #838, #860).** The inline path (`_ship_sync`) runs the `ship()` FSM transition **and** the inline `execute_ship` inside a *single* `transaction.atomic()` block. A `ShipExecutor.run()` exception — a `git push` precondition failure surfaces as a `CommandFailedError` — then rolls the `ship()` advance back, so the ship is all-or-nothing: either pushed + PR opened + FSM advanced, or the FSM is left untouched (safely re-runnable from `REVIEWED`). The exception is caught and returned as a structured `ShipExecuted` (`ok=False`, real cause in `detail`); pre-#838 it propagated unhandled, committing a partial `SHIPPED` (no push/PR) and crashing the `manage.py` subprocess so the CLI wrapper surfaced only an opaque `rc=1` with the real cause lost. `ShipExecutor.run()` also has *non-raising* precondition exits (`no code host configured`, `no worktree on ticket`, `branch … already merged into base`) that return `RunnerResult(ok=False)`; `execute_ship` then returns a normal `{"ok": False}` dict — no exception. #838 only treated an exception as the rollback trigger, so pre-#860 those structured failures still committed the same partial `SHIPPED`. #860 closes that residual path: a failing `execute_ship` result is re-raised inside the atomic block as `_ShipExecutionError` carrying the real `detail`, so both the raised and the `ok=False` paths share one rollback + structured-surfacing path. This is the synchronous analogue of the worker enqueue pattern's "state change and queued work land atomically" invariant — the async path achieves it via `on_commit`, the sync path via the shared transaction. The block's `on_commit` enqueue fires only on commit (success); `execute_ship`'s own state guard makes a later worker pickup a no-op.
 
 **Auto-scheduling:** each phase transition leads to the next-phase task in a fresh session (bias-free evaluation). `start()` schedules coding indirectly — the provision worker calls `schedule_coding()` once worktrees exist. The remaining auto-schedule edges are direct:
