+++ src/teatree/core/management/commands/pr.py
@@ -9,8 +9,6 @@ returns the PR URL once the worker completes.
 import re
 from typing import TypedDict, cast
 
-from django.db import transaction
-from django_fsm import TransitionNotAllowed
 from django_typer.management import TyperCommand, command
 
 from teatree import visual_qa
@@ -22,6 +20,13 @@ from teatree.core.management.commands._pr_preview import (
     ShipDryRun,
     ship_dry_run,
     validate_pr_metadata,
+)
+from teatree.core.management.commands._ship_exec import (
+    ShipEnqueued,
+    ShipExecuted,
+    ShippingGateFailure,
+    _enqueue_ship,
+    _ship_sync,
 )
 from teatree.core.management.commands._ship_fsm import reconcile_fsm_for_ship
 from teatree.core.models import Session, Ticket, Worktree
@@ -58,27 +63,10 @@ class NoCommitsAheadError(TypedDict):
     base: str
 
 
-class ShipEnqueued(TypedDict):
-    ticket_id: int
-    state: str
-    queued: bool
-    warning: str
-
-
-class ShipExecuted(TypedDict):
-    ticket_id: int
-    state: str
-    synced: bool
-    ok: bool
-    detail: str
-
-
-class ShippingGateFailure(TypedDict):
-    allowed: bool
-    error: str
-    missing: list[str]
-    hint: str
-
+# ShipEnqueued / ShipExecuted / ShippingGateFailure and the ship-execution
+# helpers live in ``_ship_exec`` (extracted by concern, re-exported above)
+# so external importers of ``pr.ShipExecuted`` etc. keep working and
+# ``pr.py`` stays within the module-health LOC bar.
 
 # EnsurePrResult lives in ``_ensure_pr`` (re-exported above) so external
 # importers of ``pr.EnsurePrResult`` keep working after the ensure-pr split.
@@ -177,120 +165,6 @@ def _check_shipping_gate(ticket: Ticket) -> ShippingGateFailure | None:
     # source of truth so ``ship()`` (source [REVIEWED, SHIPPED]) is legal.
     reconcile_fsm_for_ship(ticket)
     return None
-
-
-def _do_ship_transition(ticket: Ticket, title: str) -> ShippingGateFailure | None:
-    """Run the ``ship()`` FSM transition; return a gate failure or ``None``.
-
-    Invariant (#694): ``pr create`` never raises a raw
-    ``TransitionNotAllowed``. Since #748 the ``--skip-validation`` path
-    runs ``reconcile_fsm_for_ship`` too (it is the user-authorized
-    attestation substitute, so the FSM follows the authorization), so
-    ``ship()`` is normally legal here; this ``try`` remains the backstop
-    for any residual illegal hop (e.g. a state the reconcile no-ops past)
-    so the failure is reported as the same structured shape the gate-fail
-    path returns rather than raised. ``ship()`` schedules
-    ``execute_ship.enqueue`` via ``transaction.on_commit``.
-    """
-    try:
-        with transaction.atomic():
-            if title:
-                # #800 N3: canonical locked RMW (was an unlocked
-                # whole-extra overwrite — no select_for_update/re-read —
-                # racing the ship worker's pr_urls write).
-                ticket.merge_extra(set_keys={"pr_title_override": title})
-            ticket.ship()
-            ticket.save()
-    except TransitionNotAllowed:
-        return ShippingGateFailure(
-            allowed=False,
-            error=f"Cannot ship from state '{ticket.state}': FSM not in REVIEWED.",
-            missing=[],
-            hint="Drop --skip-validation so the gate can reconcile the FSM, or record the missing phases.",
-        )
-    return None
-
-
-def _enqueue_ship(ticket: Ticket, title: str) -> ShipEnqueued | ShippingGateFailure:
-    """Async ship: enqueue ``execute_ship`` and warn it needs a worker.
-
-    The push + PR are NOT performed here — they run in ``execute_ship``,
-    which only fires when a worker drains the django-tasks queue. In a
-    no-worker context (e.g. an interactive ``uv run`` invocation) the ship
-    silently never completes; the explicit ``warning`` makes that visible
-    instead of looking like a successful ship (#708). Use ``--sync`` to
-    push + open the PR inline in this process.
-    """
-    failure = _do_ship_transition(ticket, title)
-    if failure is not None:
-        return failure
-    return ShipEnqueued(
-        ticket_id=int(ticket.pk),
-        state=str(ticket.state),
-        queued=True,
-        warning=(
-            "Ship was QUEUED, not performed. The branch push and PR creation "
-            "run in the `execute_ship` task and will NOT complete until a "
-            "worker drains the queue (`t3 <overlay> tasks work-next-sdk`). "
-            "Re-run with `--sync` to push and open the PR inline now."
-        ),
-    )
-
-
-def _ship_sync(ticket: Ticket, title: str) -> ShipExecuted | ShippingGateFailure:
-    """Synchronous ship: run ``execute_ship`` inline in this process (#708).
-
-    Atomicity (#838): the FSM ``ship()`` transition and the inline
-    ``execute_ship`` run inside a **single** ``transaction.atomic()``
-    block. Pre-#838 ``_do_ship_transition`` committed ``SHIPPED`` in its
-    own transaction and ``execute_ship.call()`` ran afterwards in a
-    separate one; a ``ShipExecutor.run()`` exception (a ``git push``
-    precondition failure surfaces as ``CommandFailedError``) then left
-    ``Ticket.state == SHIPPED`` with no push and no PR — a partial state
-    the operator could not safely re-run. Sharing one transaction makes
-    the exception roll the FSM advance back, so the ship is all-or-nothing:
-    either pushed + PR opened + FSM advanced, or the FSM is untouched.
-
-    Error surfacing (#838): a ``ShipExecutor.run()`` exception is caught
-    here and returned as a structured ``ShipExecuted`` (``ok=False`` with
-    the real error in ``detail``). Pre-#838 the exception propagated
-    unhandled, crashing the ``manage.py`` subprocess so the CLI wrapper
-    only surfaced a bare ``rc=1`` with the real cause lost.
-
-    The ``on_commit`` enqueue scheduled by ``ship()`` only fires when the
-    block commits (success); ``execute_ship`` is idempotent (re-checks
-    state under ``select_for_update``) so a worker later picking it up is
-    a safe no-op (state is no longer SHIPPED).
-    """
-    from teatree.core.tasks import execute_ship  # noqa: PLC0415
-
-    try:
-        with transaction.atomic():
-            failure = _do_ship_transition(ticket, title)
-            if failure is not None:
-                return failure
-            result = execute_ship.call(int(ticket.pk))
-    except Exception as exc:  # noqa: BLE001
-        # Atomicity: the surrounding ``transaction.atomic()`` already
-        # rolled the ``ship()`` advance back, so the FSM is NOT left in a
-        # partial SHIPPED state. Surface the real cause instead of letting
-        # it crash the subprocess into an opaque wrapper-only ``rc=1``.
-        ticket.refresh_from_db()
-        return ShipExecuted(
-            ticket_id=int(ticket.pk),
-            state=str(ticket.state),
-            synced=True,
-            ok=False,
-            detail=str(exc),
-        )
-    ticket.refresh_from_db()
-    return ShipExecuted(
-        ticket_id=int(ticket.pk),
-        state=str(ticket.state),
-        synced=True,
-        ok=bool(result.get("ok", False)),
-        detail=str(result.get("detail", "")),
-    )
 
 
 def _resolve_base_url(worktree: Worktree | None) -> str:
